+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All important changes to `fiisoft-phinx` will be documented in this file
 
+## 1.6.0
+
+Added file app.php as a handy way to run commands directly (with local configuration provided by user).
+
+## 1.5.0
+
+Four new commands has been added. They are:
+* mark
+* unmark
+* remove 
+* cleanup
+
+## 1.4.0
+
+* added new methods to PhinxAbstractMigration and PhinxCommonTrait
+* added handy, shortened documentation as a comment to default template for migrations
+* removed two defaults from PhinxConfig
+
 ## 1.3.0
 
 * added new abstract class PhinxAbstractSeed